Refactor commonality between *ItemAndLocation classes
